refactoring of the framework
