+++ springbootseleniumproject/src/main/java/com/springbootselenium/springbootseleniumproject/core/Scope/BrowserScopePostProcessor.java
@@ -1,4 +1,4 @@
-package com.springbootselenium.springbootseleniumproject.config;
+package com.springbootselenium.springbootseleniumproject.core.Scope;
 
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
