+++ springbootseleniumproject/src/main/java/com/springbootselenium/springbootseleniumproject/core/annotation/LazyAutowired.java
@@ -1,0 +1,15 @@
+package com.springbootselenium.springbootseleniumproject.core.annotation;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
+
+import java.lang.annotation.*;
+
+@Lazy
+@Autowired
+@Documented
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LazyAutowired {
+}
