+++ springbootseleniumproject/src/main/java/com/springbootselenium/springbootseleniumproject/core/annotation/Page.java
@@ -1,16 +1,17 @@
-package com.springbootselenium.springbootseleniumproject.annotation;
+package com.springbootselenium.springbootseleniumproject.core.annotation;
 
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
 import java.lang.annotation.*;
 
+
 @Lazy
-@Configuration
+@Component
+@Scope("prototype")
 @Documented
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LazyConfiguration {
+public @interface Page {
 }
