+++ springbootseleniumproject/src/main/java/com/springbootselenium/springbootseleniumproject/core/annotation/PageFragment.java
@@ -1,4 +1,4 @@
-package com.springbootselenium.springbootseleniumproject.annotation;
+package com.springbootselenium.springbootseleniumproject.core.annotation;
 
 import org.springframework.context.annotation.Lazy;
 import org.springframework.context.annotation.Scope;
@@ -6,12 +6,11 @@ import org.springframework.stereotype.Component;
 
 import java.lang.annotation.*;
 
-
 @Lazy
 @Component
 @Scope("prototype")
 @Documented
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Page {
+public @interface PageFragment {
 }
