+++ springbootseleniumproject/src/main/java/com/springbootselenium/springbootseleniumproject/core/annotation/ThreadScopeBean.java
@@ -1,0 +1,16 @@
+package com.springbootselenium.springbootseleniumproject.core.annotation;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.context.annotation.Scope;
+
+import java.lang.annotation.*;
+
+@Bean
+@Scope("browserScope")
+@Documented
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ThreadScopeBean {
+}
