+++ springbootseleniumproject/src/main/java/com/springbootselenium/springbootseleniumproject/core/config/RemoteWebDriverConfig.java
@@ -1,19 +1,13 @@
-package com.springbootselenium.springbootseleniumproject.config;
+package com.springbootselenium.springbootseleniumproject.core.config;
 
-import com.springbootselenium.springbootseleniumproject.annotation.LazyConfiguration;
-import io.github.bonigarcia.wdm.WebDriverManager;
+import com.springbootselenium.springbootseleniumproject.core.annotation.LazyConfiguration;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.remote.DesiredCapabilities;
 import org.openqa.selenium.remote.RemoteWebDriver;
-import org.openqa.selenium.support.ui.WebDriverWait;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.context.annotation.Profile;
 
 import java.net.URL;
