+++ springbootseleniumproject/src/main/java/com/springbootselenium/springbootseleniumproject/page/google/GooglePage.java
@@ -1,11 +1,9 @@
 package com.springbootselenium.springbootseleniumproject.page.google;
 
-import com.springbootselenium.springbootseleniumproject.annotation.Page;
+import com.springbootselenium.springbootseleniumproject.core.annotation.Page;
 import com.springbootselenium.springbootseleniumproject.page.Base;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Lazy;
-import org.springframework.stereotype.Component;
 
 @Page
 public class GooglePage extends Base {
