+++ springbootseleniumproject/src/main/java/com/springbootselenium/springbootseleniumproject/page/google/SearchComponent.java
@@ -1,14 +1,10 @@
 package com.springbootselenium.springbootseleniumproject.page.google;
 
-import com.springbootselenium.springbootseleniumproject.annotation.PageFragment;
+import com.springbootselenium.springbootseleniumproject.core.annotation.PageFragment;
 import com.springbootselenium.springbootseleniumproject.page.Base;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.springframework.context.annotation.Lazy;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
 
 @PageFragment
 public class SearchComponent extends Base {
